+++ src/CMakeLists.txt
@@ -546,8 +546,6 @@ list( APPEND SOURCES
       export/ExportMixerDialog.h
       export/ExportOptionsHandler.cpp
       export/ExportOptionsHandler.h
-      export/ExportProgressUI.cpp
-      export/ExportProgressUI.h
 
       # Import
