+++ src/cloud/audiocom/ShareAudioDialog.cpp
@@ -43,7 +43,7 @@
 #include "CodeConversions.h"
 
 #include "Export.h"
-#include "export/ExportProgressUI.h"
+#include "ExportProgressUI.h"
 #include "ExportUtils.h"
 #include "AccessibleLinksFormatter.h"
 #include "ExportPluginRegistry.h"
